refactor(ui): split MCPSection into row/editor components and a servers hook
- settings/settingsIcons.tsx: shared SVG glyphs (also Skills' plug/download/
  check for the next step)
- settings/settingsApi.ts: shared apiSend POST/PUT/DELETE helper
- settings/mcpApi.ts: MCPList type + fetchServers
- settings/mcpRowText.ts (+tests): statusTitle / targetLine
- settings/MCPEditorCard.tsx: JSON editor card (exports MCPEditorState)
- settings/MCPServerListItem.tsx: per-server row with editor slot
- settings/useMCPServers.ts: list state + toggle/trust/delete actions
MCPSection drops to layout + editor wiring. Markup and testids unchanged.
