+++ external/ui/src/ui/settings/settingsApi.ts
@@ -1,0 +1,22 @@
+/** Shared JSON POST/PUT/DELETE helper for the /coddy settings-tab APIs. */
+export async function apiSend(
+  path: string,
+  method: "POST" | "PUT" | "DELETE",
+  body?: unknown,
+): Promise<{ ok: boolean; error?: string }> {
+  const init: RequestInit = { method };
+  if (body !== undefined) {
+    init.headers = { "Content-Type": "application/json" };
+    init.body = JSON.stringify(body);
+  }
+  const res = await fetch(path, init);
+  if (!res.ok) {
+    try {
+      const j = (await res.json()) as { error?: { message?: string } };
+      return { ok: false, error: j.error?.message || `HTTP ${res.status}` };
+    } catch {
+      return { ok: false, error: `HTTP ${res.status}` };
+    }
+  }
+  return { ok: true };
+}
